% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{all.equal.bGWAS}
\alias{all.equal.bGWAS}
\title{Equality test for bGWAS objects}
\usage{
\method{all}{equal.bGWAS}(target, current, ...)
}
\arguments{
\item{target}{an object of class bGWAS}

\item{current}{an object of class bGWAS}

\item{...}{further arguments}
}
\value{
all.equal
}
\description{
Equality test for bGWAS objects
}
