% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{availableStudies}
\alias{availableStudies}
\title{availableStudies}
\usage{
availableStudies(verbose = F)
}
\arguments{
\item{verbose}{boolean, default = FALSE}
}
\value{
Data.Frame of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics
}
