% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS.R
\name{bGWAS}
\alias{bGWAS}
\title{bGWAS - main function}
\usage{
bGWAS(name, GWAS_PofO, GWAS_additive, n_permutations = 1000,
  save_files = TRUE, verbose = TRUE)
}
\arguments{
\item{name}{The name of the analysis (character)}

\item{save_files}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}

\item{noise}{(to be removed)}

\item{sign_method}{The method used to identify significant SNPs, should be \code{"p"} for
p-value or \code{"fdr"} for false discovery rate, \code{default="p"} (character)}

\item{sign_thresh}{The threshold used to identify significant SNPs, \code{default="5e-8"}
(numeric)}

\item{use_permutations}{A logical indicating if BF p-values should be estimated using the permutation
approach,  \code{default=FALSE}}

\item{res_pruning_dist}{The distance used for pruning results (in Kb), should be between 10 and 1000,
(if set to NULL, no pruning is done), \code{default=500} (numeric)}

\item{res_pruning_LD}{The LD threshold used for pruning results, should be between 0 and 1
(if 0, distance-based pruning is used), \code{default=0} (numeric)}
}
\description{
Performs a Bayesian GWAS from Summary Statistics ... work in progress to apply the
approach to Parent of Origin (PofO) results
}
