% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Prior.R
\name{compute_prior}
\alias{compute_prior}
\title{Compute Prior}
\usage{
compute_prior(selected_studies, MR_ZMatrix, All_ZMatrix, save_files = FALSE,
  verbose = FALSE)
}
\arguments{
\item{selected_studies}{data table}

\item{MR_ZMatrix}{data table}

\item{All_ZMatrix}{data table}

\item{save_files}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\description{
From a list of significant studies (from \code{identify_StudiesMR()}), the pruned Z-matrix
of MR instruments (from \code{makeMR_ZMatrix()}) and the full Z-matrix of all SNPs for these
significant studies (from \code{makeFull_ZMatrix()}), compute the prior
}
