% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{excludeStudies}
\alias{excludeStudies}
\title{excludeStudies}
\usage{
excludeStudies(ListOfStudies = NULL, Files = NULL, Traits = NULL,
  Consortia = NULL, verbose = F)
}
\arguments{
\item{ListOfStudies}{ID of the studies, by default, use all the studies in availableStudies()}

\item{verbose}{boolean}

\item{File}{list of file names}

\item{Trait}{list of trait}

\item{Consortium}{vector, list of consortium}
}
\value{
List of studies that meet the criteria
}
\description{
excludeStudies
}
\examples{
  AllStudies = availableStudies()
  listTraits()
  MyStudies = exludeStudies(Traits=c("Smoking", "Personality"))
  nrow(AllStudies)
  nrow(AllStudies[AllStudies$ID == MyStudies, ])
}
