% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{extract_MRcoeffs_bGWAS}
\alias{extract_MRcoeffs_bGWAS}
\title{Extract MR coefficients from bGWAS results}
\usage{
extract_MRcoeffs_bGWAS(obj)
}
\arguments{
\item{obj}{an object of class bGWAS created using \code{\link{bGWAS}()}}
}
\value{
a data.frame containing the MR coefficients (1 estimate using
all chromosomes + 22 estimates with 1 chromosome masked)
}
\description{
Extracts MR coefficients (multivariate genome-wide and per-chromosome estimates)
}
