% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{get_RSquared_bGWAS}
\alias{get_RSquared_bGWAS}
\title{Get squared correlation between observed and prior effects from bGWAS results}
\usage{
get_RSquared_bGWAS(obj, SNPs = "all")
}
\arguments{
\item{obj}{an object of class bGWAS created using \code{\link{bGWAS}()}}

\item{SNPs, }{"all" / "moderate" / "instruments"}
}
\value{
a squared correlation
}
\description{
Returns squared correlation between observed and prior effects,
for different subsets of SNPs (all, the ones having at least a moderate effects - p-value < 0.001 -,
MR instruments)
}
