% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{heatmap_bGWAS}
\alias{heatmap_bGWAS}
\title{Heatmap of SNP effects on prior traits from bGWAS results}
\usage{
heatmap_bGWAS(obj, save_file = F, file_name = NULL)
}
\arguments{
\item{obj}{an object of class bGWAS created using \code{\link{bGWAS}()}}

\item{save_file}{A logical indicating if the graphic should be saved,
\code{default=FALSE}, graphic will be displayed on the on-screen device}

\item{file_name}{The name of the file saved (is \code{save_file} is \code{TRUE})
\code{default=NULL}, will used NameOfYourAnalysis_Heatmap.png}
}
\value{
a Heatmap
}
\description{
Creates a heatmap of SNP effects on prior traits
}
