% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_StudiesMR.R
\name{identify_StudiesMR}
\alias{identify_StudiesMR}
\title{Identify studies for prior using multivariate MR}
\usage{
identify_StudiesMR(ZMatrix, saveFiles = FALSE, verbose = FALSE)
}
\arguments{
\item{ZMatrix}{The pruned Z-Matrix of strong instrument or the path of the file
containing it (data table or character)}

\item{saveFiles}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\value{
Log file + list of studies significant + create files (list them) if saveFiles=T
}
\description{
From a pruned Z-matrix of strong MR instruments, performs multivariate MR
}
