% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{list_consortia}
\alias{list_consortia}
\title{Get consortia from studies that can be used to build the prior}
\usage{
list_consortia(verbose = F)
}
\value{
List of consortia
}
\description{
Get list of consortia with available GWAS summary statistics
}
