% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{list_files}
\alias{list_files}
\title{Get file names of studies that can be used to build the prior}
\usage{
list_files(IDs = NULL, verbose = F)
}
\arguments{
\item{IDs}{The IDs of the files to be listed, by default=NULL, list all files (numeric)}
}
\value{
List of files
}
\description{
Get list of studies with available GWAS summary statistics
}
