% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{list_priorGWASs}
\alias{list_priorGWASs}
\title{Get list of studies that can be used to build the prior}
\usage{
list_priorGWASs(verbose = F)
}
\arguments{
\item{verbose}{boolean, default = FALSE}
}
\value{
Data.Frame of details for all available studies
}
\description{
Get list of studies with available GWAS summary statistics
}
