% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{listConsortia}
\alias{listConsortia}
\title{listConsortia}
\usage{
listConsortia(verbose = F)
}
\value{
List of consortia
}
\description{
Get list of consortia with available GWAS summary statistics
}
