% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{listFiles}
\alias{listFiles}
\title{Get file names of studies that can be used to build the prior}
\usage{
listFiles(IDs = NULL, verbose = F)
}
\value{
List of files
}
\description{
Get list of studies with available GWAS summary statistics
}
