% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{listTraits}
\alias{listTraits}
\title{Get traits from studies that can be used to build the prior}
\usage{
listTraits(verbose = F)
}
\value{
List of traits
}
\description{
Get list of traits with available GWAS summary statistics
}
