% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFull_ZMatrix.R
\name{makeFull_ZMatrix}
\alias{makeFull_ZMatrix}
\title{Create full Z-matrix for prior computation}
\usage{
makeFull_ZMatrix(Studies = NULL, GWAS, ZMatrices = "~/ZMatrices",
  saveFiles = F, verbose = F)
}
\arguments{
\item{Studies}{The IDs of the significant studies selected by \code{identify_StudiesMR()}
(numeric vector)}

\item{GWAS}{The path to the conventionnal GWAS of interest, the ID of the GWAS from the
list of studies availables (prior GWASs), or a data.frame (character, numeric or data.frame)}

\item{ZMatrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}

\item{saveFiles}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\value{
An object containing Log file and pruned Z-Matrix of MR instrument (data table) + create a file if saveFiles=T
}
\description{
From a list of significant studies, create the full Z-matrix that can be used to compute
the prior.
}
