% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{manhattan_plot_bGWAS}
\alias{manhattan_plot_bGWAS}
\title{Manhattan Plot from bGWAS results}
\usage{
manhattan_plot_bGWAS(obj, save_file = F, file_name = NULL,
  annotate = T, SNPs = NULL, results = "BF")
}
\arguments{
\item{obj}{an object of class bGWAS created using \code{\link{bGWAS}()}}

\item{save_file}{A logical indicating if the graphic should be saved,
\code{default=FALSE}, graphic will be displayed on the on-screen device}

\item{file_name}{The name of the file saved (is \code{save_file} is \code{TRUE})
\code{default=NULL}, will used NameOfYourAnalysis_ManhattanPlot.png}

\item{annotate}{A logical indicating if the significant SNPs identified in the
analysis should be annotated on the plot, \code{default=TRUE}
If your results are not pruned or if you have a high number of significant SNPs,
be aware that \code{annotate=TRUE} might decrease readability of the figure. You
could define a set of SNPs to annotate using \code{SNPs}.}

\item{SNPs}{A data.frame containing the SNPs (rsid) to annotate in the first column, and 
optionnally the text that should be plotted in the second column, and the color
in the third column, \code{default=NULL}, only evaluated if \code{annotate=TRUE}.}

\item{results, }{"BF" / "posterior" / "direct",  \code{default="BF"}}
}
\value{
a Manhattan Plot
}
\description{
Creates a Manhattan Plot from bGWAS results (for performance, only SNPs with 
p-value or FDR < 0.05 are plotted)
}
\details{
If \code{results = "BF"}, BF p-values / fdr-values will be used. \cr
If \code{results = "direct"}, direct effect p-values / fdr-values will be used. \cr
If \code{results = "posterior"}, posterior effect p-values / fdr-values will be used. \cr
}
