% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{print.bGWAS}
\alias{print.bGWAS}
\title{Print a bGWAS object}
\usage{
\method{print}{bGWAS}(x, ...)
}
\arguments{
\item{x}{an object of class bGWAS}

\item{...}{further arguments}
}
\value{
print
}
\description{
Print a bGWAS object
}
