% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{print_log_bGWAS}
\alias{print_log_bGWAS}
\title{Print log from bGWAS results}
\usage{
print_log_bGWAS(obj)
}
\arguments{
\item{obj}{an object of class bGWAS created using \code{\link{bGWAS}()}}
}
\description{
Prints the log (everything that is printed during a \code{bGWAS} analysis)
with \code{verbose=TRUE})
}
