% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ZMatrix.R
\name{prune_ZMatrix}
\alias{prune_ZMatrix}
\title{Prune a Z-Matrix}
\usage{
prune_ZMatrix(ZMatrix, prune.dist = 5e+05, r2.limit = 1.1, verbose = F)
}
\arguments{
\item{ZMatrix}{A subset of the prior GWASs Z-Matrix}

\item{prune.dist}{The distance for pruning}

\item{r2.limit}{The r2 threshold if LD-pruning, \code{default=1.1}, only distance pruning
is performed}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\description{
Prune a Z-Matrix
}
\details{
To perform LD-pruning, need binary files !!!
}
