% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_BFandP.R
\name{request_BFandP}
\alias{request_BFandP}
\title{Calculate Bayes Factor and empirical p-values from observed Z-scores and priors}
\usage{
request_BFandP(Prior, save_files = F, verbose = F)
}
\arguments{
\item{Prior}{from the function makePrior()}

\item{save_files}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\value{
Log + data.table containing rs-chr-pos-alt-ref-obs-fit-se-...
}
\description{
From a pruned Z-matrix of strong MR instruments, performs multivariate MR
}
