% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{select_priorGWASs}
\alias{select_priorGWASs}
\title{select_priorGWASs}
\usage{
select_priorGWASs(include_names = NULL, include_traits = NULL,
  includeConsortia = NULL, exclude_names = NULL, exclude_traits = NULL,
  excludeConsortia = NULL, Z_matrices = "~/ZMatrices/", verbose = F)
}
\arguments{
\item{include_names}{list of prior GWASs names}

\item{include_traits}{list of trait}

\item{exclude_names}{list of file names}

\item{exclude_traits}{list of trait}

\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}

\item{verbose}{boolean, default = FALSE}

\item{include_consortium}{vector, list of consortium ### TO BE DONE?}

\item{exclude_consortium}{vector, list of consortium ### TO BE DONE?}
}
\value{
IDs of studies that meet the criteria
}
\description{
Allow the quick selection of a subset of studies for prior based on 3 criteria. First, include all the prior GWASs
specified (if all including parameters are NULL, include all studies), and then remove all the prior GWASs specified
(if all excluding parameters are NULL, keep all studies included at the step before)
}
\examples{
  AllStudies = list_priorGWASs()
  list_names()
  MyStudies = select_priorGWASs(include_traits=c("Heart Rate", "Body mass index", "Smoking"))
  #AllStudies[AllStudies$ID \%in\% MyStudies, ]
}
