% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_classFunctions.R
\name{all.equal.bGWAS}
\alias{all.equal.bGWAS}
\title{Equality test for bGWAS objects}
\usage{
\method{all}{equal.bGWAS}(obj1, obj2)
}
\arguments{
\item{obj1}{an object of class bGWAS}

\item{obj2}{an object of class bGWAS}
}
\value{
all.equal
}
\description{
Equality test for bGWAS objects
}
