% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS.R
\name{bGWAS}
\alias{bGWAS}
\title{bGWAS - Main Function}
\usage{
bGWAS(Name, GWAS, ExludeFromPrior = NULL, verbose = F)
}
\arguments{
\item{Name}{character, name of your analysis}

\item{GWAS}{character, path to your conventionnal GWAS of interest}

\item{ExludeFromPrior}{vector, list of studies to exclude when creating the prior, see XXX, by default=NULL}

\item{verbose}{boolean}

\item{out.path}{character, path to the outputs, by default, current working dictory}

\item{keepFiles}{?}
}
\value{
Value Resultat
}
\description{
Performs a bayesian GWAS
}
