% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS.R
\name{bGWAS}
\alias{bGWAS}
\title{bGWAS}
\usage{
bGWAS(Name, GWAS, ZMatrices = NULL, PriorStudies = NULL,
  ListOfSNPs = NULL, MRthreshold = 1e-05, SignMethod = "p",
  SignThresh = 5 * 1e-07, pruneRes = FALSE, saveFiles = FALSE,
  verbose = FALSE)
}
\arguments{
\item{Name}{The name of the analysis (character)}

\item{GWAS}{The path to the conventionnal GWAS of interest or ID of the GWAS from the
list of studies availables (prior GWASs) (character or numeric)}

\item{ZMatrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}

\item{PriorStudies}{The IDs of prior GWASs to use for the analysis, \code{default=NULL},
will include all the prior GWASs available (numeric vector)}

\item{ListOfSNPs, }{The path to a file containing the rsids of the SNPs to use ,
\code{default=NULL}, will use all the SNPs in common between prior GWASs and the
conventionnal GWAS (character)
      # NOT IMPLEMENTED YET}

\item{MRthreshold}{The threshold used to select strong instruments for MR, should be lower
than 1e-5, \code{default=1e-5} (numeric)}

\item{SignMethod}{The method used to identify significant SNPs, should be \code{"p"} for
p-value or \code{"fdr"} for false discovery rate, \code{default="p"} (character)}

\item{SignThresh}{The threshold used to identify significant SNPs, \code{default="5.10e-8"}
(numeric)}

\item{pruneRes}{A logical indicating if the results should be pruned (by distance, 500kb),
\code{default=FALSE}}

\item{saveFiles}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\value{
An data.table containing the significant SNPs (...) + log file saved in working directory
if saveFiles=T, additional files containing the Z-Matrices used (for MR / to build the prior), the
coefficents from the regression, the BF and p-values estimated for all SNPs / subset for significant
SNPs are created
## DESCRIBE THE FILES
}
\description{
Performs a bayesian GWAS from Summary Statistics, using publicly available results
to calculate the prior effects of the SNPs and compare it to observed z-scores
}
\examples{
# Permorm bGWAS, using a conventional GWAS from the list of prior GWASs
MyGWAS = 1
listFiles(MyGWAS)
\dontrun{
A = bGWAS(Name = "Test_UsingGWASfromList",
         GWAS = MyGWAS
         verbose=T)
         }
# Permorm bGWAS, using a small conventional GWAS included in data and selecting a subset of
# studies for the prior
MyGWAS = system.file("Data/SmallGWAS_Pilling2017.csv", package="bGWAS")
MyStudies = selectStudies(includeTraits=c("Type 2 diabetes", "Smoking"),
                         includeFiles=c("jointGwasMc_HDL.txt.gz","jointGwasMc_LDL.txt.gz"))
\dontrun{
B = bGWAS(Name = "Test_UsingSmallGWAS",
         GWAS = MyGWAS,
         PriorStudies=MyStudies,
         verbose=T)
         }
}
