% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS.R
\name{bGWAS}
\alias{bGWAS}
\title{bGWAS - main function}
\usage{
bGWAS(name, GWAS, Z_matrices = "~/ZMatrices/", prior_studies = NULL,
  SNPs_list = NULL, MR_threshold = 1e-05, MR_pruning_dist = 500,
  MR_pruning_LD = 0.2, MR_shrinkage = 1e-05, prior_shrinkage = 1e-05,
  stop_after_prior = FALSE, sign_method = "p", sign_thresh = 5e-08,
  use_permutations = FALSE, res_pruning_dist = 500, res_pruning_LD = 0,
  save_files = FALSE, verbose = TRUE)
}
\arguments{
\item{name}{The name of the analysis (character)}

\item{GWAS}{The path to the conventional GWAS of interest, the ID of the GWAS from the
list of studies available (prior GWASs), or a data.frame (character, numeric or data.frame)}

\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}

\item{prior_studies}{The IDs of prior GWASs to use for the analysis, \code{default=NULL},
will include all the prior GWASs available (numeric vector)}

\item{SNPs_list, }{The path to a file containing the rsids of the SNPs to use ,
\code{default=NULL}, will use all the SNPs in common between prior GWASs and the
conventionnal GWAS (character)
      # NOT IMPLEMENTED YET}

\item{MR_threshold}{The threshold used to select strong instruments for MR, should be lower
than 1e-5, \code{default=1e-5} (numeric)}

\item{MR_pruning_dist}{The distance used for pruning MR instruments (in Kb), should be between 10 and 1000,
\code{default=500} (numeric)}

\item{MR_pruning_LD}{The LD threshold used for pruning MR instruments, should be between 0 and 1
(if>=1, distance-based pruning is used), \code{default=0.2} (numeric)}

\item{MR_shrinkage}{The p-value threshold used for shrinkage before performing MR, should be between
\code{MR_threshold} and 1, \code{default=1e-5} (numeric)}

\item{prior_shrinkage}{The p-value threshold used for shrinkage before calculating the prior,
should be between \code{MR_threshold} and 1, \code{default=1e-5} (numeric)}

\item{sign_method}{The method used to identify significant SNPs, should be \code{"p"} for
p-value or \code{"fdr"} for false discovery rate, \code{default="p"} (character)}

\item{sign_thresh}{The threshold used to identify significant SNPs, \code{default="5e-8"}
(numeric)}

\item{res_pruning_dist}{The distance used for pruning results (in Kb), should be between 10 and 1000,
(if set to NULL, no pruning is done), \code{default=500} (numeric)}

\item{res_pruning_LD}{The LD threshold used for pruning results, should be between 0 and 1
(if>=1, distance-based pruning is used), \code{default=0} (numeric)}

\item{save_files}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}

\item{sign_method}{The method used to identify significant SNPs, should be \code{"p"} for}
}
\value{
\code{bGWAS}() returns an object of class "bGWAS". \cr
Additionnaly, if \code{save_files=T}, several files are created... \cr
... in your working directory :
\itemize{
\item  "\code{name}.log" - log file
}
... in the folder \code{./name/} :
\itemize{
\item "PriorGWASs.tsv" - contains Prior GWASs information
(general info + status (used/removed) + MR coefficients)
\item "CoefficientsByChromosome.csv" - contains the MR estimates when masking the focal
chromosome (22 coefficients / Prior GWAS)
\item "PriorBFp.csv" - contains BF and p-values estimated for all SNPs
\item "SignificantSNPs.csv" - contains BF and p-values estimated for a subset of SNPs
}
}
\description{
Performs a Bayesian GWAS from Summary Statistics, using publicly available results
to calculate the prior effects of the SNPs and compare it to observed z-scores
}
\details{
\code{Name} and \code{GWAS} are required arguments.
If \code{GWAS} is a path to a file (regular or .gz) or a data.frame, it should contain the following
columns : \cr
SNPID (rs numbers) should be : \code{rs}, \code{rsid}, \code{snp}, \code{snpid}, \code{rnpid} \cr
A1 should be : \code{a1}, \code{alt}, \code{alts} \cr
A2 should be : \code{a2}, \code{a0}, \code{ref} \cr
Z should be : \code{z}, \code{Z}, \code{zscore} \cr
If Z is not present, it can be calculated from BETA and SE, in this case, a temporary
gzipped file is created and removed after the analysis. \cr
BETA should be : \code{b}, \code{beta}, \code{beta1} \cr
SE should be : \code{se}, \code{std} \cr
Z-Matrix files, containing Z-scores for all prior GWASs should be downloaded separately
and stored in \code{"~/ZMatrices"} or in the folder specified with the argument
\code{Z_matrices}. \cr
## MORE INFO NEEDED HERE / HOW TO DOWNLOAD ID (see GitHub README)

Use \code{\link{list_priorGWASs}()} to see all the prior GWASs available.
Using one of them as your conventionnal GWAS (argument \code{GWAS} = numeric ID) will automatically
remove it from the list of prior GWASs used to build the prior.

Use \link{select_priorGWASs}() to automatically select the studies to
be included/excluded when building the prior (argument \code{prior_studies}).

Be careful, in the results, all the GWAS (conventional + prior) are aligned with UK10K
data for the analysis (some alleles might be swapped when comparing with the initial data)
}
\examples{
# Permorm bGWAS, using a conventional GWAS from the list of prior GWASs
MyGWAS = 5
list_priorGWASs(MyGWAS)
\dontrun{
A = bGWAS(name = "Test_UsingGWASfromPriorGWASs",
         GWAS = MyGWAS)
         }

# Permorm bGWAS, using a small conventional GWAS included in data (file) and selecting a subset of
# studies for the prior
MyGWAS = system.file("Data/SmallGWAS_Pilling2017.csv", package="bGWAS")
MyStudies = select_priorGWASs(include_traits=c("Type 2 diabetes", "Smoking"),
                         include_files=c("jointGwasMc_HDL.txt.gz","jointGwasMc_LDL.txt.gz"))
\dontrun{
B = bGWAS(name = "Test_UsingSmallGWAS",
          GWAS = MyGWAS,
          prior_studies=MyStudies,
          verbose=F)
         }

#'# Permorm bGWAS, using a small conventional GWAS included in data (data.frame) and selecting a subset
# of studies for the prior
\dontrun{
data("SmallGWAS_Pilling2017")
C = bGWAS(name="Test_UsingSmallDataFrame",
          GWAS = SmallGWAS_Pilling2017,
          prior_studies=MyStudies,
          save_files=T)
          }

# Note that B and C are using the same data (stored differently) and give the same results.
}
