% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SignificantSNPs.R
\name{get_SignificantSNPs}
\alias{get_SignificantSNPs}
\title{Identify significant SNPs}
\usage{
get_SignificantSNPs(Prior, SignMethod = "p", SignThresh = 5e-08,
  pruneRes = T, saveFiles = F, verbose = F)
}
\arguments{
\item{Prior}{from the function request_BFandP()}

\item{SignMethod}{The method used to identify significant SNPs, should be \code{"p"} for
p-value or \code{"fdr"} for false discovery rate, \code{default="p"} (character)}

\item{SignThresh}{The threshold used to identify significant SNPs, \code{default="5e-8"}
(numeric)}

\item{pruneRes}{A logical indicating if the results should be pruned (by distance, 500kb),
\code{default=FALSE}}

\item{saveFiles}{A logical indicating if the results should be saved as files,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=FALSE}}
}
\value{
Log + data.table containing rs-chr-pos-alt-ref-obs-fit-se-...
}
\description{
From BFs and p-values, identify significant SNPs given a specified threshold and prune them
if necessary
}
