% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_PriorGWASs.R
\name{list_files}
\alias{list_files}
\title{List prior GWASs files}
\usage{
list_files(IDs = NULL, Z_matrices = "~/ZMatrices/")
}
\arguments{
\item{IDs}{the IDs of the studies to print, \code{default="~/ZMatrices/"} will list
all of them (numeric),}

\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}
}
\value{
List of files
}
\description{
Lists the filenames of the prior GWASs
}
