% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_PriorGWASs.R
\name{list_priorGWASs}
\alias{list_priorGWASs}
\title{List prior GWASs}
\usage{
list_priorGWASs(IDs = NULL, Z_matrices = "~/ZMatrices/")
}
\arguments{
\item{IDs}{the IDs of the studies to print, \code{default="~/ZMatrices/"} will list
all of them (numeric),}

\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}
}
\value{
a \code{tibble} containing prior GWASs information
}
\description{
Lists the studies that can be used as prior GWASs
}
