% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_PriorGWASs.R
\name{list_traits}
\alias{list_traits}
\title{List prior GWASs traits
Lists the traits of the prior GWASs}
\usage{
list_traits(Z_matrices = "~/ZMatrices/")
}
\arguments{
\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}
}
\value{
List of traits
}
\description{
List prior GWASs traits
Lists the traits of the prior GWASs
}
