% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGWAS_PriorGWASs.R
\name{select_priorGWASs}
\alias{select_priorGWASs}
\title{Select prior GWASs}
\usage{
select_priorGWASs(include_files = NULL, include_traits = NULL,
  exclude_files = NULL, exclude_traits = NULL,
  Z_matrices = "~/ZMatrices/", verbose = F)
}
\arguments{
\item{include_files}{list of file names (see \code{\link{list_files}()}) (character)}

\item{include_traits}{list of trait (see \code{\link{list_traits}()}) (character)}

\item{exclude_files}{list of file names (see \code{\link{list_files}()}) (character)}

\item{exclude_traits}{list of trait (see \code{\link{list_traits}()}) (character)}

\item{Z_matrices}{The path to the folder containing Z-Matrices, \code{default="~/ZMatrices/"}
(character)}

\item{verbose}{boolean, default = FALSE}
}
\value{
IDs (numeric) of studies that meet the criteria
}
\description{
Allow the quick selection of a subset of prior GWASs based on 2 criteria. First, include all the files
specified (if all including parameters are NULL, include all studies), and then remove all the files specified
(if all excluding parameters are NULL, keep all studies included at the step before)
}
\examples{
  \dontrun{
  AllStudies = list_priorGWASs()
  list_traits()
  MyStudies = select_priorGWASs(include_traits=c("Heart Rate", "Body Mass Index", "Smoking"))
  AllStudies[AllStudies$ID \%in\% MyStudies, c("ID", "Name", "Trait", "File")]}
}
