% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{select_priorGWASs}
\alias{select_priorGWASs}
\title{select_priorGWASs}
\usage{
select_priorGWASs(include_files = NULL, include_traits = NULL,
  includeConsortia = NULL, exclude_files = NULL, exclude_traits = NULL,
  excludeConsortia = NULL, verbose = F)
}
\arguments{
\item{include_files}{list of file names}

\item{include_traits}{list of trait}

\item{exclude_files}{list of file names}

\item{exclude_traits}{list of trait}

\item{verbose}{boolean}

\item{include_consortium}{vector, list of consortium ### TO BE DONE}

\item{exclude_consortium}{vector, list of consortium ### TO BE DONE}
}
\value{
IDs of studies that meet the criteria
}
\description{
Allow the quick selection of a subset of studies for prior based on 3 criteria. First, include all the files
specified (if all including parameters are NULL, include all studies), and then remove all the files specified
(if all excluding parameters are NULL, keep all studies included at the step before)
}
\examples{
  AllStudies = list_priorGWASs()
  list_traits()
  MyStudies = select_priorGWASs(include_traits=c("Heart Rate", "Body mass index", "Smoking"))
  #AllStudies[AllStudies$ID \%in\% MyStudies, ]
}
