% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudiesForPrior.R
\name{selectStudies}
\alias{selectStudies}
\title{selectStudies}
\usage{
selectStudies(ListOfStudies = NULL, Files = NULL, Traits = NULL,
  Consortia = NULL, verbose = F)
}
\arguments{
\item{ListOfStudies}{ID of the studies, by default, use all the studies in availableStudies()}

\item{verbose}{boolean}

\item{File}{list of file names}

\item{Trait}{list of trait}

\item{Consortium}{vector, list of consortium}
}
\value{
List of studies that meet the criteria
}
\description{
Allow the quick selection of a subset of studies for prior based on 3 criteria
}
\examples{
  AllStudies = availableStudies()
  listTraits()
  MyStudies = selectStudies(Traits=c("Heart Rate", "Body mass index", "Smoking"))
  AllStudies[AllStudies$ID == MyStudies, ]
}
